+++ packages/fiori-app-sub-generator/test/unit/fiori-app-generator/end.test.ts
@@ -1,5 +1,5 @@
 import { runPostGenerationTasks } from '../../../src/fiori-app-generator/end';
-import { generateLaunchConfig, runHooks, t } from '../../../src/utils';
+import { DEFAULT_POST_APP_GEN_COMMAND, generateLaunchConfig, runHooks, t } from '../../../src/utils';
 import type { BackendSystem } from '@sap-ux/store';
 import { getService } from '@sap-ux/store';
 import { PLATFORMS, ApiHubType, EventName, State } from '../../../src/types';
@@ -63,7 +63,8 @@ describe('runPostGenerationTasks', () => {
         const project = {
             targetFolder: '/path/to/project',
             name: 'testProject',
-            flpAppId: 'testAppId'
+            flpAppId: 'testAppId',
+            enableVirtualEndpoints: false
         };
 
         await runPostGenerationTasks({ service, project }, fs, logger, vscode, appWizard);
@@ -75,7 +76,8 @@ describe('runPostGenerationTasks', () => {
                 flpAppId: project.flpAppId,
                 sapClientParam: 'sap-client=100',
                 odataVersion: service.odataVersion,
-                datasourceType: DatasourceType.sapSystem
+                datasourceType: DatasourceType.sapSystem,
+                enableVirtualEndpoints: project.enableVirtualEndpoints
             },
             fs,
             vscode,
@@ -177,7 +179,7 @@ describe('runPostGenerationTasks', () => {
             {
                 hookParameters: { fsPath: projectPath },
                 vscodeInstance: vscode,
-                options: { followUpCommand: undefined }
+                options: { command: undefined }
             },
             logger
         );