+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.1.11
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.35.7
+
+## 0.1.10
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.35.6
+
 ## 0.1.9
 
 ### Patch Changes