+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,94 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 0.1.31
+
+### Patch Changes
+
+-   b9675bb: Improve Fiori Tools UI Texts
+-   Updated dependencies [b9675bb]
+    -   @sap-ux/deploy-config-generator-shared@0.1.14
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.28
+    -   @sap-ux/fiori-generator-shared@0.12.13
+
+## 0.1.30
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/ui5-config@0.28.3
+    -   @sap-ux/abap-deploy-config-writer@0.1.13
+    -   @sap-ux/fiori-generator-shared@0.12.12
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.27
+    -   @sap-ux/deploy-config-generator-shared@0.1.13
+
+## 0.1.29
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.26
+-   @sap-ux/deploy-config-generator-shared@0.1.12
+-   @sap-ux/ui5-config@0.28.2
+-   @sap-ux/abap-deploy-config-writer@0.1.12
+-   @sap-ux/fiori-generator-shared@0.12.11
+
+## 0.1.28
+
+### Patch Changes
+
+-   Updated dependencies [4fef16a]
+    -   @sap-ux/fiori-generator-shared@0.12.10
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.25
+    -   @sap-ux/deploy-config-generator-shared@0.1.11
+
+## 0.1.27
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.24
+-   @sap-ux/deploy-config-generator-shared@0.1.10
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/abap-deploy-config-writer@0.1.11
+-   @sap-ux/fiori-generator-shared@0.12.9
+
+## 0.1.26
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.23
+-   @sap-ux/deploy-config-generator-shared@0.1.9
+-   @sap-ux/abap-deploy-config-writer@0.1.10
+
+## 0.1.25
+
+### Patch Changes
+
+-   Updated dependencies [5a88606]
+    -   @sap-ux/abap-deploy-config-inquirer@1.4.22
+
+## 0.1.24
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.21
+-   @sap-ux/deploy-config-generator-shared@0.1.9
+-   @sap-ux/abap-deploy-config-writer@0.1.9
+
+## 0.1.23
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.4.20
+-   @sap-ux/deploy-config-generator-shared@0.1.9
+-   @sap-ux/abap-deploy-config-writer@0.1.8
+
+## 0.1.22
+
+### Patch Changes
+
+-   Updated dependencies [3ea811a]
+    -   @sap-ux/abap-deploy-config-writer@0.1.7
+
 ## 0.1.21
 
 ### Patch Changes