+++ packages/annotation-generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/annotation-generator
 
+## 0.3.42
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/fiori-annotation-api@0.6.5
+
+## 0.3.41
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/fiori-annotation-api@0.6.4
+
 ## 0.3.40
 
 ### Patch Changes