+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,81 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.25
+
+### Patch Changes
+
+-   b9675bb: Improve Fiori Tools UI Texts
+-   Updated dependencies [b9675bb]
+-   Updated dependencies [b9675bb]
+    -   @sap-ux/deploy-config-generator-shared@0.1.14
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.21
+    -   @sap-ux/cf-deploy-config-writer@0.3.8
+    -   @sap-ux/fiori-generator-shared@0.12.13
+    -   @sap-ux/inquirer-common@0.7.18
+
+## 0.2.24
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/cf-deploy-config-writer@0.3.7
+    -   @sap-ux/fiori-generator-shared@0.12.12
+    -   @sap-ux/deploy-config-generator-shared@0.1.13
+    -   @sap-ux/inquirer-common@0.7.17
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.20
+
+## 0.2.23
+
+### Patch Changes
+
+-   Updated dependencies [48af01a]
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.19
+
+## 0.2.22
+
+### Patch Changes
+
+-   @sap-ux/deploy-config-generator-shared@0.1.12
+-   @sap-ux/fiori-generator-shared@0.12.11
+-   @sap-ux/inquirer-common@0.7.16
+-   @sap-ux/cf-deploy-config-inquirer@0.3.18
+
+## 0.2.21
+
+### Patch Changes
+
+-   Updated dependencies [4fef16a]
+    -   @sap-ux/fiori-generator-shared@0.12.10
+    -   @sap-ux/deploy-config-generator-shared@0.1.11
+    -   @sap-ux/inquirer-common@0.7.15
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.17
+
+## 0.2.20
+
+### Patch Changes
+
+-   @sap-ux/deploy-config-generator-shared@0.1.10
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/cf-deploy-config-writer@0.3.6
+-   @sap-ux/fiori-generator-shared@0.12.9
+-   @sap-ux/inquirer-common@0.7.14
+-   @sap-ux/cf-deploy-config-inquirer@0.3.16
+
+## 0.2.19
+
+### Patch Changes
+
+-   Updated dependencies [ff6ff97]
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.15
+
+## 0.2.18
+
+### Patch Changes
+
+-   Updated dependencies [3ea811a]
+    -   @sap-ux/cf-deploy-config-writer@0.3.5
+
 ## 0.2.17
 
 ### Patch Changes