+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.1.14
+
+### Patch Changes
+
+-   b9675bb: Improve Fiori Tools UI Texts
+-   Updated dependencies [b9675bb]
+    -   @sap-ux/fiori-generator-shared@0.12.13
+
+## 0.1.13
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.12.12
+
+## 0.1.12
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.12.11
+
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies [4fef16a]
+    -   @sap-ux/fiori-generator-shared@0.12.10
+
+## 0.1.10
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.12.9
+
 ## 0.1.9
 
 ### Patch Changes