+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.35.7
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/fiori-annotation-api@0.6.5
+
+## 0.35.6
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/fiori-annotation-api@0.6.4
+
 ## 0.35.5
 
 ### Patch Changes