+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.6.5
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/cds-odata-annotation-converter@0.6.3
+
+## 0.6.4
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/cds-odata-annotation-converter@0.6.3
+
 ## 0.6.3
 
 ### Patch Changes