+++ packages/ui5-library-reference-inquirer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/ui5-library-reference-inquirer
 
+## 0.4.18
+
+### Patch Changes
+
+-   b9675bb: Improve Fiori Tools UI Texts
+-   Updated dependencies [b9675bb]
+    -   @sap-ux/inquirer-common@0.7.18
+
+## 0.4.17
+
+### Patch Changes
+
+-   Updated dependencies [f75b89d]
+    -   @sap-ux/project-access@1.30.4
+    -   @sap-ux/inquirer-common@0.7.17
+
+## 0.4.16
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.16
+
+## 0.4.15
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.15
+
+## 0.4.14
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/inquirer-common@0.7.14
+
 ## 0.4.13
 
 ### Patch Changes